:truck: Renaming the Commands package to PublicCommands
