+++ src/main/java/be/isservers/hmb/CommandManager.java
@@ -4,13 +4,13 @@ import be.isservers.hmb.command.IPrivateCommand;
 import be.isservers.hmb.command.IPublicCommand;
 import be.isservers.hmb.command.PrivateCommandContext;
 import be.isservers.hmb.command.PublicCommandContext;
-import be.isservers.hmb.command.commands.HelpCommand;
-import be.isservers.hmb.command.commands.JokeCommand;
-import be.isservers.hmb.command.commands.MemeCommand;
-import be.isservers.hmb.command.commands.PingCommand;
-import be.isservers.hmb.command.commands.admin.ClearChannelCommand;
-import be.isservers.hmb.command.commands.admin.SetPrefixCommand;
-import be.isservers.hmb.command.commands.music.*;
+import be.isservers.hmb.command.publicCommands.HelpCommand;
+import be.isservers.hmb.command.publicCommands.JokeCommand;
+import be.isservers.hmb.command.publicCommands.MemeCommand;
+import be.isservers.hmb.command.publicCommands.PingCommand;
+import be.isservers.hmb.command.publicCommands.admin.ClearChannelCommand;
+import be.isservers.hmb.command.publicCommands.admin.SetPrefixCommand;
+import be.isservers.hmb.command.publicCommands.music.*;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
 import net.dv8tion.jda.api.events.message.priv.PrivateMessageReceivedEvent;
