+++ src/main/java/be/isservers/hmb/command/publicCommands/music/SkipCommand.java
@@ -1,4 +1,4 @@
-package be.isservers.hmb.command.commands.music;
+package be.isservers.hmb.command.publicCommands.music;
 
 import be.isservers.hmb.command.IPublicCommand;
 import be.isservers.hmb.command.PublicCommandContext;
